Refactor helper classes into different files
